Fixed where in reader tests to check for time slices
Previously had been testing that the time axis would be chopped to the
desired data range (and frequency) in the low-level private reader
routines (specific to each machine/database). However, this
transformation can be better handled once the data is placed in a
DataArray object, meaning it should occur in the higher-level abstract
reader class. I've updated the tests to reflect this.