Corrections so that SXR fitting works
Note that for cameras T and H it seems to struggle to adequately fit
the peak of the data. That is when I am just fitting to the data of
one camera. When I fit to the data of all three I have the issue that
camera H has smaller luminosities than the other two and the fitting
seems to favour lower luminosities. As such, the fit gets dragged down
towards camera H. Does it need to be calibrated, perhaps?
Alternatively, there might be an issue with how I'm handling
transverse skew; I don't think that's been tested properly.