Implemented generic parts of reading bremsstrahlung data
I made some small changes to the interface for fetching data from the
database. This will allow it to work with any sensors that have
multiple lines of sight. Overall, the interface is now more like that
for radiation quantities (to the point where I could almost use the
same generic code).