Fixed error in SURF reader
This error happened to show up while the standard test-suite was being
run for this branch, even though no changes were made which affected
the SURF reader. It seems it just took this long for Hypothesis to
happen upon some failing input. The problem was that using an asterisk
character for the instrument name could end up confusing the SURF reader.