Add unit test for radiation inversion
The tests starts by creating an emissivity profile of the type
returned by InvertRadiation operators. This is used to work out the
fluxes that would be observed along various lines of sight. These are
then passed to the operator and the results compared to the expected
emissivity profile generated at the start. The solver does not
converge to as accurate an answer as I'd like, with particular issues
dealing with asymmetry.