Create dummy-PPF data for testing PPF reader #25
PPF data is proprietary, meaning we can not commit it to the
repository. This posed a challenge for my approach to testing the PPF
reader, as that relied on having a sample of some PPF data in a pickle
file. However, I realised that I just need data that is the _right
shape_, and the actual values could be overwritten with random
numbers. I've modified the script which fetches the data to make that
possible and committed a sample of this dummy data to the
repository. We will now be able to run these unit tests in CI (which
should give a nice boost to code coverage!).